requirements.txt
argcomplete==1.11.1
asgiref==3.2.5
automium==0.2.6
automium-web==0.1.1
awscli==1.18.23
boto3==1.12.23
botocore==1.15.23
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Django==3.0.4
djangorestframework==3.11.0
docutils==0.15.2
durationpy==0.5
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
hjson==3.0.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
kappa==0.6.0
legacy==0.1.6
MarkupSafe==1.1.1
pip-tools==4.5.1
placebo==0.9.0
pyasn1==0.4.8
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
smmap==3.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
tqdm==4.43.0
troposphere==2.6.0
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
